\name{Simple Course Management System}
\alias{simplecms}
\title{Simple Course Management System}
\usage{
Invoke the Addin from the Addins menu in RStudio.
}
\description{
Useful on an RStudio Server Pro instance, for instructors who wish to
distribute handout files to students, collect and return homework files,
and monitor in-class work that's happening on the server.
}
